Fix exp() to expf()
[ROCm/rocRAND commit: 1a6f24722e6959ee04f87bf7cc4112657ba55f59]

+++ library/src/rng/distribution/log_normal.hpp
@@ -116,8 +116,8 @@ struct mrg_log_normal_distribution<float>
     float2 operator()(const unsigned int x, const unsigned int y)
     {
         float2 v = rocrand_device::detail::mrg_normal_distribution2(x, y);
-        v.x = exp(mean + (stddev * v.x));
-        v.y = exp(mean + (stddev * v.y));
+        v.x = expf(mean + (stddev * v.x));
+        v.y = expf(mean + (stddev * v.y));
         return v;
     }
 };
